requirements.txt
Babel==2.15.0
blinker==1.8.2
click==8.1.7
docxcompose==1.4.0
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
python-docx==1.1.2
setuptools==72.1.0
six==1.16.0
typing_extensions==4.12.2
Werkzeug==3.0.3
boto3